+++ src/shard.h
@@ -19,8 +19,8 @@ typedef void (*shard_complete_cb)(uint8_t id, void *arg);
 
 typedef struct shard {
 	struct shard_state {
-		uint32_t sent;
-		uint32_t tried_sent;
+		uint64_t sent;
+		uint64_t tried_sent;
 		uint32_t blacklisted;
 		uint32_t whitelisted;
 		uint32_t failures;
