+++ src/state.h
@@ -172,7 +172,7 @@ struct state_recv {
 	// valid responses NOT classified as "success"
 	uint32_t failure_total;
 	// valid responses that passed the filter
-	uint32_t filter_success;
+	uint64_t filter_success;
 	// how many packets did we receive that were marked as being the first
 	// fragment in a stream
 	uint32_t ip_fragments;
